version with test

+++ src/main/java/com/example/at_springboot/model/service/DepartamentoService.java
@@ -2,6 +2,7 @@ package com.example.at_springboot.model.service;
 
 import com.example.at_springboot.model.dto.DepartamentoDTO;
 import com.example.at_springboot.model.entity.Departamento;
+import com.example.at_springboot.model.exceptions.ValidacaoException;
 import com.example.at_springboot.model.repository.DepartamentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,15 +15,28 @@ public class DepartamentoService {
     @Autowired
     DepartamentoRepository departamentoRepository;
 
+    @Autowired
+    UtilitarioValidacao utilitarioValidacao;
+
     public List<Departamento> ListarTodos() {
         return departamentoRepository.findAll();
     }
 
     public Departamento BuscarPorId(Long id) {
+        if (!utilitarioValidacao.verificarId(id)) {
+            Departamento departamento = new Departamento();
+            return departamento;
+        }
         return departamentoRepository.findById(id) .orElseThrow(() -> new RuntimeException("Departamento não encontrado com ID: " + id));
     }
 
     public Departamento Salvar(Departamento departamento) {
+        if (!utilitarioValidacao.verificarVazioString(departamento.getNome()) ||
+                !utilitarioValidacao.verificarVazioString(departamento.getLocal()) ) {
+            departamento = new Departamento();
+            return departamento;
+        }
+
         return departamentoRepository.save(departamento);
     }
 
@@ -32,12 +46,24 @@ public class DepartamentoService {
     }
 
     public boolean Excluir(Long id) {
+        System.out.println(id);
+        System.out.println(id);
+        System.out.println(id);
+        
+        if (!utilitarioValidacao.verificarId(id)) {
+            throw new ValidacaoException("id incorreto");
+
+        }
+
         try{
             departamentoRepository.deleteById(id);
+
+
             return true;
         }
         catch (Exception e) {
-            return false;
+            throw new ValidacaoException("id incorreto");
+
         }
     }
 }
